requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
pytz==2024.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.136.0
gspread==6.1.2